Rename 'common' to 'intern'
