+++ lib/intern.js
